Cleaned-up code for building in Windows and Linux

+++ src/Common/Parameters.h
@@ -175,8 +175,8 @@ struct envStochParams {
 #if BUTTERFLYDISP
 	bool fromFile;
 #endif
-	float ac; float std;
-	float locExtProb;
+	double ac; double std; 
+	double locExtProb;
 };
 
 class paramStoch {
@@ -196,9 +196,9 @@ private:
 	bool local;				// applied locally (if not, application is global)
 	bool inK;					// in carrying capacity (if not, in growth rate)
 	bool localExt;		// local extinction applied
-	float ac;					// temporal autocorrelation coefficient
-	float std;				// amplitude of fluctuations: sampled from N(0,std)
-	float locExtProb;	// local extinction probability
+	double ac;					// temporal autocorrelation coefficient		
+	double std;				// amplitude of fluctuations: sampled from N(0,std)
+	double locExtProb;	// local extinction probability
 };
 
 //---------------------------------------------------------------------------
